Move/rename plugins
Right idea, just premature
