+++ htdocs/modules/publisher/class/plugin/index.html
