+++ htdocs/modules/publisher/class/plugin/mainmenu.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace XoopsModules\Publisher\Plugin;
+//namespace XoopsModules\Publisher\Plugin;
 
 /*
  You may not change or alter any portion of this comment or credits
@@ -12,36 +12,33 @@ namespace XoopsModules\Publisher\Plugin;
  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  */
 
-use Criteria;
-use CriteriaCompo;
-use WaitingPluginInterface;
-use XoopsModules\Publisher\Helper;
+use MainmenuPluginInterface;
+use Xoops;
 
 /**
  * @copyright       The XOOPS Project http://sourceforge.net/projects/xoops/
  * @license         http://www.fsf.org/copyleft/gpl.html GNU public license
  * @author          trabis <lusopoemas@gmail.com>
- * @author          Laurent JEN (aka DuGris)
  */
-class WaitingPlugin implements WaitingPluginInterface
+class PublisherMainmenuPlugin implements MainmenuPluginInterface
 {
     /**
      * @return array
      */
-    public function waiting()
+    public function mainmenu()
     {
-        $helper = Helper::getInstance();
-        $ret = [];
-        $criteria = new CriteriaCompo();
-        $criteria->add(new Criteria('status', 1));
-        $count = $helper->getItemHandler()->getCount($criteria);
-        if ($count) {
-            $ret[] = [
-                'count' => $count,
-                'name' => _MI_PUBLISHER_WAITING,
-                'link' => $helper->url('admin/item.php'),
-            ];
+        $helper = Xoops::getModuleHelper(\basename(\dirname(\dirname(__DIR__))));
+        $subMenu = [];
+        // Prevent wasting resources
+        if ($helper->isCurrentModule()) {
+            //Todo: Implement submenu;
         }
+
+        $ret[] = [
+            'name' => $helper->getModule()->getVar('name'),
+            'link' => $helper->url(),
+            'subMenu' => $subMenu,
+        ];
 
         return $ret;
     }
